fix(dashboard): app UX redesign phase 1
- Split app-detail-workspace.tsx monolith (1193 LOC) into:
  - app-detail-workspace.tsx (shell)
  - app-overview-tab.tsx
  - app-activity-tab.tsx
  - app-settings-tab.tsx
  - app-detail-primitives.tsx
  - app-receipt-components.tsx
  - app-types.ts
- Redesign tab bar: compact, icon-only mobile, ARIA roles, keyboard nav
- Simplify activity filters: collapse behind toggle, remove dead batch UI
- Remove checkboxes from receipt rows (no bulk action existed)
- Fix scrollbar hell: remove nested scrollbars, overscroll-contain
- Fix review layout grid squeeze with minmax(0,1fr)
- Stack scope cards and action buttons on mobile
- Calm colors: stopped uses blue not attention
