+++ dashboard/src/apps/app-types.ts
@@ -1,0 +1,7 @@
+export type AppStatus = "active" | "needs_setup" | "observed" | "unknown";
+
+export type TabKey = "overview" | "activity" | "settings";
+
+export type AppFilter = "all" | "pending" | "allowed" | "blocked";
+
+export type AppTimeFilter = "all" | "today" | "week";
